refactor: Antag Datums (#3157)

+++ code/datums/status_effects/status_effect.dm
@@ -23,7 +23,7 @@
 		LAZYADD(owner.status_effects, src)
 	if(!owner || !on_apply())
 		qdel(src)
-		return
+		return FALSE
 	if(duration != -1)
 		duration = world.time + duration
 	tick_interval = world.time + tick_interval
@@ -55,17 +55,22 @@
 		tick()
 		tick_interval = world.time + initial(tick_interval)
 	if(duration != -1 && duration < world.time)
+		on_timeout()
 		qdel(src)
 
 /datum/status_effect/proc/on_apply() //Called whenever the buff is applied; returning FALSE will cause it to autoremove itself.
 	return TRUE
 /datum/status_effect/proc/tick() //Called every tick.
 /datum/status_effect/proc/on_remove() //Called whenever the buff expires or is removed; do note that at the point this is called, it is out of the owner's status_effects but owner is not yet null
+/datum/status_effect/proc/on_timeout()  // Called specifically whenever the status effect expires.
 /datum/status_effect/proc/be_replaced() //Called instead of on_remove when a status effect is replaced by itself or when a status effect with on_remove_on_mob_delete = FALSE has its mob deleted
 	owner.clear_alert(id)
 	LAZYREMOVE(owner.status_effects, src)
 	owner = null
 	qdel(src)
+
+/datum/status_effect/proc/before_remove() //! Called before being removed; returning FALSE will cancel removal
+	return TRUE
 
 /datum/status_effect/proc/refresh()
 	var/original_duration = initial(duration)
@@ -102,6 +107,8 @@
 
 /mob/living/proc/apply_status_effect(effect, ...) //applies a given status effect to this mob, returning the effect if it was successful
 	. = FALSE
+	if(QDELETED(src))
+		return
 	var/datum/status_effect/S1 = effect
 	LAZYINITLIST(status_effects)
 	for(var/datum/status_effect/S in status_effects)
@@ -118,15 +125,20 @@
 	S1 = new effect(arguments)
 	. = S1
 
-/mob/living/proc/remove_status_effect(effect) //removes all of a given status effect from this mob, returning TRUE if at least one was removed
+
+/// Removes all of a given status effect from this mob, returning TRUE if at least one was removed
+/mob/living/proc/remove_status_effect(effect, ...)
 	. = FALSE
+	var/list/arguments = args.Copy(2)
 	if(status_effects)
 		var/datum/status_effect/S1 = effect
 		for(var/datum/status_effect/S in status_effects)
-			if(initial(S1.id) == S.id)
+			if(initial(S1.id) == S.id && S.before_remove(arglist(arguments)))
 				qdel(S)
 				. = TRUE
 
+
+/// Returns the effect if the mob calling the proc owns the given status effect, or null otherwise
 /mob/living/proc/has_status_effect(effect) //returns the effect if the mob calling the proc owns the given status effect
 	. = FALSE
 	if(status_effects)
@@ -134,6 +146,16 @@
 		for(var/datum/status_effect/S in status_effects)
 			if(initial(S1.id) == S.id)
 				return S
+
+
+/// Returns the effect if the mob calling the proc owns the given status effect, but checks by type.
+/mob/living/proc/has_status_effect_type(effect)
+	if(!length(status_effects))
+		return
+	for(var/datum/status_effect/S in status_effects)
+		if(istype(S, effect))
+			return S
+
 
 /mob/living/proc/has_status_effect_list(effect) //returns a list of effects with matching IDs that the mod owns; use for effects there can be multiple of
 	. = list()
@@ -159,6 +181,7 @@
 	var/max_stacks //stacks cannot exceed this amount
 	var/consumed_on_threshold = TRUE //if status should be removed once threshold is crossed
 	var/threshold_crossed = FALSE //set to true once the threshold is crossed, false once it falls back below
+	var/reset_ticks_on_stack = FALSE //resets the current tick timer if a stack is gained
 	var/overlay_file
 	var/underlay_file
 	var/overlay_state // states in .dmi must be given a name followed by a number which corresponds to a number of stacks. put the state name without the number in these state vars
@@ -201,6 +224,8 @@
 	owner.cut_overlay(status_overlay)
 	owner.underlays -= status_underlay
 	stacks += stacks_added
+	if(reset_ticks_on_stack)
+		tick_interval = world.time + initial(tick_interval)
 	if(stacks > 0)
 		if(stacks >= stack_threshold && !threshold_crossed) //threshold_crossed check prevents threshold effect from occuring if changing from above threshold to still above threshold
 			threshold_crossed = TRUE
@@ -220,8 +245,9 @@
 		qdel(src) //deletes status if stacks fall under one
 
 /datum/status_effect/stacking/on_creation(mob/living/new_owner, stacks_to_apply)
-	..()
-	add_stacks(stacks_to_apply)
+	. = ..()
+	if(.)
+		add_stacks(stacks_to_apply)
 
 /datum/status_effect/stacking/on_apply()
 	if(!can_have_status())
@@ -247,10 +273,12 @@
 	QDEL_NULL(status_overlay)
 	return ..()
 
+
 /// Status effect from multiple sources, when all sources are removed, so is the effect
 /datum/status_effect/grouped
 	status_type = STATUS_EFFECT_MULTIPLE //! Adds itself to sources and destroys itself if one exists already, there are never multiple
 	var/list/sources = list()
+
 
 /datum/status_effect/grouped/on_creation(mob/living/new_owner, source)
 	var/datum/status_effect/grouped/existing = new_owner.has_status_effect(type)
@@ -262,9 +290,11 @@
 		sources |= source
 		return ..()
 
-/datum/status_effect/grouped/on_remove(source)
+
+/datum/status_effect/grouped/before_remove(source)
 	sources -= source
 	return !length(sources)
+
 
 /**
  * # Transient Status Effect (basetype)
@@ -278,10 +308,12 @@
 	/// How much strength left before expiring? time in deciseconds.
 	var/strength = 0
 
+
 /datum/status_effect/transient/on_creation(mob/living/new_owner, set_duration)
 	if(isnum(set_duration))
 		strength = set_duration
 	. = ..()
+
 
 /datum/status_effect/transient/tick()
 	if(QDELETED(src) || QDELETED(owner))
@@ -291,6 +323,7 @@
 	if(strength <= 0)
 		qdel(src)
 		return FALSE
+
 
 /**
  * Returns how much strength should be adjusted per tick.
